backend/requirements.txt
# FastAPI
fastapi
uvicorn[standard]

# Environment
python-dotenv

# Data Processing
pandas
numpy

# Machine Learning
scikit-learn
joblib

# Database
pymongo

# File Upload
python-multipart

# Email
sendgrid

# Visualization
matplotlib

# CORS (included in fastapi but explicit)
starlette
